Move item chat card functions to its own mixin
